[Style/Like] 좋아요 페이지 요소들 정렬
