+++ src/app/(tools)/like/LikeList.tsx
@@ -1,50 +1,50 @@
-"use client";
+'use client';
 
-import styled from "styled-components";
-import ListItem from "@/components/list/ListItem";
-import ListNav from "@/components/list/ListNav";
-import PageSort from "@/app/(nav)/board/PageSort";
+import styled from 'styled-components';
+import ListNav from '@/components/list/ListNav';
+import List from '@/components/list/List';
+import PageSort from '@/app/(nav)/board/PageSort';
 
 const LikeListBlock = styled.div`
-  width: 80%;
-  max-width: 1200px;
+  width: 100%;
   display: flex;
   flex-direction: column;
+  align-items: center;
 `;
 
-const GaroLine = styled.div`
-  border-top: 1px solid rgba(0, 0, 0, 0.2);
-`;
-
-const ListItemBlock = styled.div`
-  //border-top: 1px solid rgba(0, 0, 0, 0.2);
-  padding: 30px 0;
-  display: grid;
-  grid-template-columns: repeat(auto-fill, minmax(250px, 1fr));
-  row-gap: 30px;
-  justify-items: center;
-`;
-
-const LikeList = () => {
+export default function LikeList() {
   return (
     <LikeListBlock>
       <ListNav />
-      <GaroLine />
-      <div className="self-end pt-5 mr-8 w-40">
-        <PageSort />
-      </div>
-      <ListItemBlock>
-        <ListItem />
-        <ListItem />
-        <ListItem />
-        <ListItem />
-        <ListItem />
-        <ListItem />
-        <ListItem />
-        <ListItem />
-      </ListItemBlock>
+      <List />
     </LikeListBlock>
   );
-};
+}
 
-export default LikeList;
+// const GaroLine = styled.div`
+//   border-top: 1px solid rgba(0, 0, 0, 0.2);
+// `;
+
+// const ListItemBlock = styled.div`
+//   //border-top: 1px solid rgba(0, 0, 0, 0.2);
+//   padding: 30px 0;
+//   display: grid;
+//   grid-template-columns: repeat(auto-fill, minmax(250px, 1fr));
+//   row-gap: 30px;
+//   justify-items: center;
+// `;
+
+// const LikeList = () => {
+//   return (
+//     <LikeListBlock>
+//       <ListNav />
+//       {/* <List /> */}
+//       {/* <div className='self-end pt-5 mr-8 w-40'> */}
+//       {/* <PageSort /> */}
+//       {/* </div> */}
+//       <List />
+//     </LikeListBlock>
+//   );
+// };
+
+// export default LikeList;
