+++ src/components/product-detail-page/pdp-right/elements/UserInfo.tsx
@@ -1,10 +1,10 @@
 'use client';
 
 import styled from 'styled-components';
-import UserPic from '@/components/userpage/userinfo/UserPic';
-import UserName from '@/components/userpage/userinfo/UserName';
-import UserTemp from '@/components/userpage/UserTemp';
-import UserAd from '@/components/userpage/userinfo/UserAd';
+import UserPic from '@/components/userpage/user/userinfo/UserPic';
+import UserName from '@/components/userpage/user/userinfo/UserName';
+import UserTemp from '@/components/userpage/user/UserTemp';
+import UserAd from '@/components/userpage/user/userinfo/UserAd';
 import Link from 'next/link';
 
 const UserInfoBox = styled.div`
