+++ src/components/userpage/user/userinfo/UserAd.tsx
@@ -1,4 +1,13 @@
-const UserAd = () => {
-  return <address>경기도 고양시</address>;
-};
-export default UserAd;
+'use client';
+
+import styled from 'styled-components';
+
+// user 주소
+const Address = styled.address`
+  font-size: 10px;
+  color: rgba(33, 33, 36, 0.5);
+`;
+
+export default function UserAd() {
+  return <Address>경기도 고양시</Address>;
+}
