+++ src/components/userpage/user/userinfo/UserName.tsx
@@ -1,4 +1,3 @@
-const UserName = () => {
-  return <p>키티키티키팈</p>;
-};
-export default UserName;
+export default function UserName() {
+  return <p className='hover:cursor-pointer'>키티키티키팈</p>;
+}
