+++ src/components/userpage/user/userinfo/UserPic.tsx
@@ -3,20 +3,26 @@
 import Image from 'next/image';
 import styled from 'styled-components';
 
+type Props = {
+  size?: number;
+};
+
 const ImageDeco = styled(Image)`
   padding: 0 10px 10px 10px;
   border-radius: 50%;
   overflow: hidden;
+  &:hover {
+    cursor: pointer;
+  }
 `;
 
-const UserPic = () => {
+export default function UserPic({ size = 150 }: Props) {
   return (
     <ImageDeco
       src='/images/for-demo/kitty.jpg'
       alt='kitty'
-      width={150}
-      height={150}
+      width={size}
+      height={size}
     />
   );
-};
-export default UserPic;
+}
