--- src/components/userpage/userinfo/UserName.tsx
@@ -1,3 +1,0 @@
-export default function UserName() {
-  return <p className='hover:cursor-pointer'>키티키티키팈</p>;
-}
